ey_project/requirements.txt
streamlit
chromadb
sentence-transformers
openai
python-dotenv
rdkit-pypi
pyvis
networkx
tqdm
fpdf2